board page first setup

+++ spring/src/main/java/pl/mkowsky/jirawannabedemo/JwtResponse.java
@@ -5,14 +5,14 @@ import java.util.List;
 public class JwtResponse {
 	private String token;
 	private String type = "Bearer";
-	private int id;
+	private Long id;
 	private String username;
 	private String email;
 	private List<String> roles;
 
 
 
-	public JwtResponse(String accessToken, int id, String username, String email, List<String> roles) {
+	public JwtResponse(String accessToken, Long id, String username, String email, List<String> roles) {
 		this.token = accessToken;
 		this.id = id;
 		this.username = username;
@@ -39,11 +39,11 @@ public class JwtResponse {
 		this.type = tokenType;
 	}
 
-	public int getId() {
+	public Long getId() {
 		return id;
 	}
 
-	public void setId(int id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
