+++ spring/src/main/java/pl/mkowsky/jirawannabedemo/model/Role.java
@@ -8,7 +8,7 @@ public class Role {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
 
     @Enumerated(EnumType.STRING)
     private ERole name;
@@ -21,11 +21,13 @@ public class Role {
         this.name = name;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+
+
+    public void setId(Long id) {
         this.id = id;
     }
