+++ spring/src/main/java/pl/mkowsky/jirawannabedemo/security/services/UserDetailsImpl.java
@@ -14,7 +14,7 @@ import java.util.stream.Collectors;
 public class UserDetailsImpl implements UserDetails {
     private static final long serialVersionUID = 1L;
 
-    private int id;
+    private Long id;
 
     private String username;
 
@@ -25,7 +25,7 @@ public class UserDetailsImpl implements UserDetails {
 
     private Collection<? extends GrantedAuthority> authorities;
 
-    public UserDetailsImpl(int id, String username, String email, String password,
+    public UserDetailsImpl(Long id, String username, String email, String password,
                            Collection<? extends GrantedAuthority> authorities) {
         this.id = id;
         this.username = username;
@@ -56,7 +56,7 @@ public class UserDetailsImpl implements UserDetails {
         return serialVersionUID;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
